Avoid very expensive $phys_functions_f90.cpp compile if SCREAM_LIB_ONLY
