+++ components/scream/src/physics/p3/tests/p3_tests.cpp
@@ -1,5 +1,6 @@
 #include "catch2/catch.hpp"
 #include "physics/p3/p3_f90.hpp"
+#include "physics/p3/p3_main_wrap.hpp"
 #include "physics/p3/p3_ic_cases.hpp"
 
 namespace {
