+++ components/scream/src/physics/shoc/CMakeLists.txt
@@ -1,12 +1,18 @@
 set(SHOC_SRCS
   shoc_f90.cpp
-  shoc_functions_f90.cpp
   shoc_ic_cases.cpp
   shoc_iso_c.f90
   shoc_iso_f.f90
   ${SCREAM_BASE_DIR}/../eam/src/physics/cam/shoc.F90
   atmosphere_macrophysics.cpp
 )
+
+if (NOT SCREAM_LIB_ONLY)
+  list(APPEND SHOC_SRCS
+    shoc_functions_f90.cpp
+    shoc_main_wrap.cpp
+    ) # Add f90 bridges needed for testing
+endif()
 
 set(SHOC_HEADERS
   shoc.hpp
