+++ components/scream/src/physics/shoc/tests/shoc_run_and_cmp.cpp
@@ -1,4 +1,4 @@
-#include "physics/shoc/shoc_f90.hpp"
+#include "physics/shoc/shoc_main_wrap.hpp"
 #include "physics/shoc/shoc_functions_f90.hpp"
 #include "physics/shoc/shoc_ic_cases.hpp"
