+++ components/scream/src/physics/shoc/tests/shoc_tests.cpp
@@ -1,6 +1,6 @@
 #include "catch2/catch.hpp"
 
-#include "physics/shoc/shoc_f90.hpp"
+#include "physics/shoc/shoc_main_wrap.hpp"
 #include "physics/shoc/shoc_ic_cases.hpp"
 
 #include "ekat/util/ekat_test_utils.hpp"
